refactor: move server/agents tests into test/server/agents
