+++ test/server/agents/codex-args.spec.ts
@@ -1,5 +1,5 @@
 import { describe, it, expect } from "vitest";
-import { buildCodexArgs } from "./codex-args.js";
+import { buildCodexArgs } from "../../../../server/../../server/agents/codex-args.js";
 
 const base = { resume: null, model: null, guiMcpUrl: null };
